move snap to utils
